fix: fixed comments on PR
