+++ src/store/modules/bot-commands/bot-commands-getters.ts
@@ -1,6 +1,6 @@
 import { GetterTree } from 'vuex'
 import { RootState } from '@/store/types'
-import { BotCommandsState } from '@/store/modules/bot-commands/types.ts'
+import { BotCommandsState } from '@/store/modules/bot-commands/types'
 
 export const getters: GetterTree<BotCommandsState, RootState> = {
   getCommandsHistory: (state) => (parentId: string) => {
