+++ src/store/modules/bot-commands/bot-commands-state.ts
@@ -1,4 +1,4 @@
-import { BotCommandsState } from '@/store/modules/bot-commands/types.ts'
+import { BotCommandsState } from '@/store/modules/bot-commands/types'
 
 export const state: BotCommandsState = {
   commands: {}
