+++ src/store/modules/bot-commands/index.ts
@@ -1,10 +1,10 @@
-import { actions } from './bot-commands-actions.ts'
-import { mutations } from './bot-commands-mutations.ts'
-import { state } from './bot-commands-state.ts'
-import { getters } from './bot-commands-getters.ts'
+import { actions } from './bot-commands-actions'
+import { mutations } from './bot-commands-mutations'
+import { state } from './bot-commands-state'
+import { getters } from './bot-commands-getters'
 import { Module } from 'vuex'
 import { RootState } from '@/store/types'
-import { BotCommandsState } from '@/store/modules/bot-commands/types.ts'
+import { BotCommandsState } from '@/store/modules/bot-commands/types'
 
 const bodCommandsModule: Module<BotCommandsState, RootState> = {
   namespaced: true,
